Add struct mod
