+++ src/bin/02_plugin_drivers/domain/customer_id.rs
@@ -1,0 +1,35 @@
+use super::traits::GenSerialData;
+
+pub struct CustomerID {
+    id: Option<String>,
+    digit: usize,
+    name: String,
+}
+
+impl CustomerID {
+    pub fn new(digit: usize) -> Self {
+        CustomerID {
+            name: "UserID".to_owned(),
+            digit,
+            id: None,
+        }
+    }
+}
+
+impl GenSerialData for CustomerID {
+    fn get_length(&mut self) -> usize {
+        self.digit
+    }
+
+    fn get_rawdata(&self) -> String {
+        self.id.clone().unwrap()
+    }
+
+    fn get_name(&self) -> String {
+        self.name.clone()
+    }
+
+    fn put_rawdata(&mut self, data: String) {
+        self.id = Some(data);
+    }
+}
