+++ src/bin/02_plugin_drivers/domain/mod.rs
@@ -1,1 +1,3 @@
 pub mod traits;
+pub mod customer_id;
+pub mod product_id;
